Finalisation API Manager
- Passage des ID à Long (Permet de le set à null car c'est un Object et non un type primitif).

- Ajout de serializable partout.

- Suppression de lien entre truck et incident avant de supprimer un truck.

- Refonte de la gestion de la récupération des données depuis le script python.
À partir du string retourne une liste avec soi un incident avec un id (récupéré depuis la base) sinon non et donc une création.
On applique la nouvelle intensité sur l'incident.
Si jamais l'intensité est à 0 alors on supprime l'incident de la base.
On envoi tous les incidents reçu au MQTT.

- Suppression du seeder

- Passage des ID du schema de base de donnée à des bigint pour mapper avec Long.

+++ src/main/java/com/projettransversal/api/Controllers/BarrackController.java
@@ -24,7 +24,7 @@ public class BarrackController {
     }
 
     @RequestMapping(value = "/barrack/{id}", method = RequestMethod.GET)
-    public Optional<Barrack> getBarrackById(@PathVariable int id) {
+    public Optional<Barrack> getBarrackById(@PathVariable Long id) {
         return _barrackService.findById(id);
     }
