+++ src/main/java/com/projettransversal/api/Controllers/IncidentController.java
@@ -17,17 +17,17 @@ public class IncidentController {
         _incidentService = incidentService;
     }
 
-    @RequestMapping(value="/incidents", method= RequestMethod.GET)
+    @RequestMapping(value = "/incidents", method = RequestMethod.GET)
     public List<Incident> getIncidents() {
         return _incidentService.findAll();
     }
 
-    @RequestMapping(value="/incident/{id}", method= RequestMethod.GET)
-    public Optional<Incident> getIncidentById(@PathVariable int id) {
+    @RequestMapping(value = "/incident/{id}", method = RequestMethod.GET)
+    public Optional<Incident> getIncidentById(@PathVariable Long id) {
         return _incidentService.findById(id);
     }
 
-    @RequestMapping(value="/incident/data", method=RequestMethod.POST)
+    @RequestMapping(value = "/incident/data", method = RequestMethod.POST)
     public void postData(@RequestBody DataRequestDTO dataRequestDTO) throws JsonProcessingException {
         this._incidentService.addData(dataRequestDTO);
     }
