+++ src/main/java/com/projettransversal/api/Repositories/BarrackRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BarrackRepository extends CrudRepository<Barrack, Integer> {
+public interface BarrackRepository extends CrudRepository<Barrack, Long> {
 
 }
