+++ src/main/java/com/projettransversal/api/Repositories/MapItemRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MapItemRepository extends CrudRepository<MapItem, Integer> {
+public interface MapItemRepository extends CrudRepository<MapItem, Long> {
 
     @Query(value = "select * from map_item where posx = ?1 and posy = ?2", nativeQuery = true)
     MapItem findByCoordinates(int x, int y);
